feat: implement authorization

+++ src/services/notes/controller/note-controller.js
@@ -1,15 +1,21 @@
-import { InvariantError, NotFoundError } from '../../../exceptions/index.js';
+import {
+  InvariantError,
+  NotFoundError,
+  AuthorizationError,
+} from '../../../exceptions/index.js';
 import response from '../../../utils/response.js';
 
 import NoteRepositories from '../repository/note-repositories.js';
 
 export const createNote = async (req, res, next) => {
   const { title, tags, body } = req.validated;
+  const { id: owner } = req.user;
 
   const note = await NoteRepositories.createNote({
     title,
     body,
     tags,
+    owner,
   });
 
   if (!note) {
@@ -27,12 +33,20 @@ export const getNotes = async (req, res) => {
   //   const note = await NoteRepositories.getNotes(title);
   //   return response(res, 200, 'success', { notes: note });
   // }
-  const notes = await NoteRepositories.getNotes();
+  const { id: owner } = req.user;
+  const notes = await NoteRepositories.getNotes(owner);
   return response(res, 200, 'success', { notes });
 };
 
 export const getNoteById = async (req, res, next) => {
   const { id } = req.params;
+
+  const { id: owner } = req.user;
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+  if (!isOwner) {
+    return next(new AuthorizationError('Anda tidak berhak mengakses resource ini'));
+  }
+
   const note = await NoteRepositories.getNoteById(id);
 
   if (!note) {
@@ -45,6 +59,13 @@ export const getNoteById = async (req, res, next) => {
 export const editNoteById = async (req, res, next) => {
   const { id } = req.params;
   const { title, tags, body } = req.validated;
+
+  const { id: owner } = req.user;
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+  if (!isOwner) {
+    return next(new AuthorizationError('Anda tidak berhak mengakses resource ini'));
+  }
+
   const note = await NoteRepositories.editNoteById({
     id,
     title,
@@ -61,6 +82,13 @@ export const editNoteById = async (req, res, next) => {
 
 export const deleteNoteById = async (req, res, next) => {
   const { id } = req.params;
+
+  const { id: owner } = req.user;
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+  if (!isOwner) {
+    return next(new AuthorizationError('Anda tidak berhak mengakses resource ini'));
+  }
+
   const deletedNote = await NoteRepositories.deleteNoteById(id);
 
   if (!deletedNote) {
